Refactor sticky calling notification on top of services branch
